+++ packages/components/src/tag/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './tagBackground';
